@@ -1,6 +1,6 @@
 import "./App.css";
 import Header from "./components/common/header/Header";
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import About from "./components/about/About";
 import CourseHome from "./components/allcourses/CourseHome";
 import Team from "./components/team/Team";
@@ -14,58 +14,82 @@ import SignUp from "./SignUp";
 
 function App() {
   return (
-    <>
-      <Router>
-        <Switch>
-          <Route exact path="/" component={Login} />
-          <Route exact path="/signup" component={SignUp} />
-          <Route exact path="/home">
-            <Header />
-            <Home />
-            <Footer />
-          </Route>
-
-          <Route exact path="/about">
-            <Header />
-            <About />
-            <Footer />
-          </Route>
-
-          <Route exact path="/courses">
-            <Header />
-            <CourseHome />
-            <Footer />
-          </Route>
-
-          <Route exact path="/team">
-            <Header />
-            <Team />
-            <Footer />
-          </Route>
-
-          <Route exact path="/realtimepair">
-            <Header />
-            <RealtimePair />
-            <Footer />
-          </Route>
-
-          <Route exact path="/journal">
-            <Header />
-            <Blog />
-            <Footer />
-          </Route>
-
-          <Route exact path="/contact">
-            <Header />
-            <Contact />
-            <Footer />
-          </Route>
-          <Route exact path="*">
-            <h1>404 Not Found</h1>
-          </Route>
-        </Switch>
-      </Router>
-    </>
+    <Router>
+      <Routes>
+        <Route path="/" element={<Login />} />
+        <Route path="/signup" element={<SignUp />} />
+        <Route
+          path="/home"
+          element={
+            <>
+              <Header />
+              <Home />
+              <Footer />
+            </>
+          }
+        />
+        <Route
+          path="/about"
+          element={
+            <>
+              <Header />
+              <About />
+              <Footer />
+            </>
+          }
+        />
+        <Route
+          path="/courses"
+          element={
+            <>
+              <Header />
+              <CourseHome />
+              <Footer />
+            </>
+          }
+        />
+        <Route
+          path="/team"
+          element={
+            <>
+              <Header />
+              <Team />
+              <Footer />
+            </>
+          }
+        />
+        <Route
+          path="/realtimepair"
+          element={
+            <>
+              <Header />
+              <RealtimePair />
+              <Footer />
+            </>
+          }
+        />
+        <Route
+          path="/journal"
+          element={
+            <>
+              <Header />
+              <Blog />
+              <Footer />
+            </>
+          }
+        />
+        <Route
+          path="/contact"
+          element={
+            <>
+              <Header />
+              <Contact />
+              <Footer />
+            </>
+          }
+        />
+      </Routes>
+    </Router>
   );
 }
 
